feat(http): add default proxy rule filtered query params
