+++ Cargo.toml
@@ -13,7 +13,7 @@ members = [
 ]
 
 [workspace.package]
-version = "1.0.31"
+version = "1.0.32"
 authors = ["The Dragonfly Developers"]
 homepage = "https://d7y.io/"
 repository = "https://github.com/dragonflyoss/client.git"
@@ -23,14 +23,14 @@ readme = "README.md"
 edition = "2021"
 
 [workspace.dependencies]
-dragonfly-client = { path = "dragonfly-client", version = "1.0.31" }
-dragonfly-client-core = { path = "dragonfly-client-core", version = "1.0.31" }
-dragonfly-client-config = { path = "dragonfly-client-config", version = "1.0.31" }
-dragonfly-client-storage = { path = "dragonfly-client-storage", version = "1.0.31" }
-dragonfly-client-backend = { path = "dragonfly-client-backend", version = "1.0.31" }
-dragonfly-client-metric = { path = "dragonfly-client-metric", version = "1.0.31" }
-dragonfly-client-util = { path = "dragonfly-client-util", version = "1.0.31" }
-dragonfly-client-init = { path = "dragonfly-client-init", version = "1.0.31" }
+dragonfly-client = { path = "dragonfly-client", version = "1.0.32" }
+dragonfly-client-core = { path = "dragonfly-client-core", version = "1.0.32" }
+dragonfly-client-config = { path = "dragonfly-client-config", version = "1.0.32" }
+dragonfly-client-storage = { path = "dragonfly-client-storage", version = "1.0.32" }
+dragonfly-client-backend = { path = "dragonfly-client-backend", version = "1.0.32" }
+dragonfly-client-metric = { path = "dragonfly-client-metric", version = "1.0.32" }
+dragonfly-client-util = { path = "dragonfly-client-util", version = "1.0.32" }
+dragonfly-client-init = { path = "dragonfly-client-init", version = "1.0.32" }
 dragonfly-api = "=2.1.78"
 thiserror = "2.0"
 futures = "0.3.31"
